requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Werkzeug==2.3.7
SQLAlchemy==2.0.21
python-dotenv==1.0.0
requests==2.26.0
dnspython==2.1.0
beautifulsoup4==4.9.3
python-whois==0.7.3
python-nmap==0.7.1
reportlab==4.0.4
Pillow==9.5.0
gunicorn==20.1.0
pytest==6.2.5
coverage==5.5
arabic-reshaper==2.1.4
python-bidi==0.4.2
email-validator==2.0.0.post2 